Starting to verify the contents

+++ Deploy.ps1
@@ -1,3 +1,4 @@
+param ([switch]$fast = $false)
 $script:scriptPath = split-path -parent $MyInvocation.MyCommand.Definition 
 cd $scriptPath
 
@@ -10,6 +11,31 @@ function test-return() {
     if ($LASTEXITCODE -ne 0) {
         write-error "Command failed with code $LASTEXITCODE"
     }
+}
+
+# Test the contents of the Vsix to make sure it has all of the appropriate
+# files 
+function test-vsixcontents() { 
+    param ([string]$vsixPath = $(throw "Need the path to the VSIX")) 
+    if (-not (test-Path $vsixPath)) {
+        write-error "Vsix doesn't exist"
+    }
+
+    # Make a folder to hold the files
+    $target = join-path ([IO.Path]::GetTempPath()) ([IO.Path]::GetRandomFileName())
+    mkdir $target | out-null
+
+    # Copy the VSIX to a file with a zip extension.  This is required for the 
+    # shell to unzip it for us. 
+    $vsixTarget = join-path $target "VsVim.zip"
+    copy $vsixPath $vsixTarget
+
+    # Unzip the file 
+    $shellApp = new-object -com shell.application
+    $source = $shellApp.NameSpace($vsixTarget)
+    $dest = $shellApp.NameSpace($target)
+    $items = $source.Items()
+    $dest.Copyhere($items, 20)
 }
 
 function build-clean() {
@@ -31,13 +57,15 @@ function build-release() {
 }
 
 # First step is to clean out all of the projects 
-write-host "Cleaning Projects"
-build-clean VsVimSpecific\VsVimDev10.csproj
-build-clean VsVimSpecific\VsVimDev11.csproj
-build-clean VsVim\VsVim.csproj
+if (-not $fast) { 
+    write-host "Cleaning Projects"
+    build-clean VsVimSpecific\VsVimDev10.csproj
+    build-clean VsVimSpecific\VsVimDev11.csproj
+    build-clean VsVim\VsVim.csproj
 
-if (test-path "VsVim\VsVim.Dev11.dll") {
-    rm VsVim\VsVim.Dev11.dll
+    if (test-path "VsVim\VsVim.Dev11.dll") {
+        rm VsVim\VsVim.Dev11.dll
+    }
 }
 
 # Next step is to build the 2012 components.  We need to get the 2010 specific DLL
@@ -50,4 +78,7 @@ copy VsVimSpecific\bin11\Release\VsVim.Dev11.dll VsVim
 
 # Now build the final output project
 build-release VsVim\VsVim.csproj
+
+$vsixPath = "VsVim\bin\Release\VsVim.vsix"
+test-vsixcontents $vsixPath
